requirements.txt
discord.py
python-dotenv
traceback2